WORKAROUND: Added greenmile query based on route date
Currently incorrect baseline start times are imported due to a lack of datetimes in the route file.

Improvements include A to B, B to A comparison for the master route sheet. 

Renaming variables to better match greenmile. Date no longer refers to the baselineStart time. Date now coorellates to GM Route route.date and baselineStart is route.baselineStartTime

TODO: Settings!!!